backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0

# Validation & settings
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# Auth
PyJWT>=2.10.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1

# HTTP client (for OAuth & AI API calls)
httpx>=0.28.0

# AI / LLM
openai>=1.30.0

# Logging
python-json-logger>=3.2.0
